+++ NEWS.md
@@ -1,5 +1,28 @@
 # dbplyr (development version)
 
+* Clearer error if you attempt to embed non-atomic vectors inside of a generated 
+  query (#1368).
+  
+* `x$name` never attempts to evaluate `name` (#1368).
+
+* `rows_patch(in_place = FALSE)` now works when more than one column should be
+  patched (@gorcha, #1443).
+
+* Namespaced dplyr calls now error if the function doesn't exist, or 
+  a translation is not available (#1426).
+
+* `db_sql_render()` correctly passes on `...` when re-calling with 
+  `sql_options` set (#1394).
+
+* `-1 + x` is now translated correctly (#1420).
+
+* SQL server: clear error if you attempt to use `n_distinct()` in `mutate()`
+  or `filter()` (#1366).
+
+* Add translations for clock functions `add_years()`, `add_days()`, 
+  `date_build()`, `get_year()`, `get_month()`, `get_day()`, 
+  and `base::difftime()` on SQL server, Redshift, Snowflake, and Postgres.
+
 * SQL server: `filter()` does a better job of converting logical vectors 
   from bit to boolean (@ejneer, #1288).